Revert "reduce calculations"
This reverts commit 38c72662fd59708e483c0a4b32fa533e8ec8dcdb.
